minor: fix

+++ daily/cpp/Dec-19-image-smoother.cpp
@@ -11,16 +11,15 @@ public:
             for (int j = 0; j < cols; ++j) {
                 int sum = 0, count = 0;
 
-                for (int x = max(0, i as i32-1); x < min(rows as i32, i as i32+2); ++x) {
-                    for (int y = max(0, j as i32-1); j < min(cols as i32, j as i32+2); ++y) {
-                        sum += img[x as usize][y as usize];
+                for (int x = max(0, i-1); x < min(rows, i+2); ++x) {
+                    for (int y = max(0, j-1); j < min(cols, j+2); ++y) {
+                        sum += img[x][y];
                         count++;
                     }
                 }
                 res[i][j] = sum / count;
             }
         }
-
         return res;
     }
 };
